# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API Service de Notification",
      "description": "Un microservice complet pour l'envoi de notifications par email, SMS et notifications push.",
      "version": "1.0.0",
      "contact": {
        "email": "support@roadtrip.com"
      },
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "servers": [
      {
        "url": "/api",
        "description": "Serveur de développement"
      }
    ],
...
    "paths": {
...
      "/users/login": {
        "post": {
          "tags": ["Auth"],
          "summary": "Connexion utilisateur",
          "description": "Authentification d'un utilisateur existant",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["email", "password"],
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email",
                      "example": "utilisateur@example.com"
                    },
                    "password": {
                      "type": "string",
                      "format": "password",
                      "example": "MotDePasse123!"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Connexion réussie",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "token": {
                        "type": "string"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "user": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Identifiants incorrects",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        }
      },
...
      "/users/forgotPassword": {
        "post": {
          "tags": ["Auth"],
          "summary": "Demande de réinitialisation de mot de passe",
          "description": "Envoie un email et/ou SMS avec un lien ou code de réinitialisation",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["email"],
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email",
                      "example": "utilisateur@example.com"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Email/SMS de réinitialisation envoyé",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "message": {
                        "type": "string",
                        "example": "Token de réinitialisation envoyé à votre email"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Utilisateur non trouvé",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        }
      },
...
      "/notifications": {
        "get": {
          "tags": ["Notifications"],
          "summary": "Récupérer toutes les notifications",
          "description": "Obtient toutes les notifications de l'utilisateur connecté",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "description": "Numéro de page pour la pagination",
              "schema": {
                "type": "integer",
                "default": 1
              }
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Nombre d'éléments par page",
              "schema": {
                "type": "integer",
                "default": 10
              }
            },
            {
              "name": "type",
              "in": "query",
              "description": "Filtrer par type de notification",
              "schema": {
                "type": "string",
                "enum": ["email", "sms", "push"]
              }
            },
            {
              "name": "status",
              "in": "query",
              "description": "Filtrer par statut de notification",
              "schema": {
                "type": "string",
                "enum": ["pending", "sent", "delivered", "failed", "read"]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Liste des notifications récupérée avec succès",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "string",
                        "example": "success"
                      },
                      "results": {
                        "type": "integer",
                        "example": 10
                      },
                      "total": {
                        "type": "integer",
                        "example": 50
                      },
                      "page": {
                        "type": "integer",
                        "example": 1
                      },
                      "pages": {
                        "type": "integer",
                        "example": 5
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "notifications": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/Notification"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Non autorisé",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          }
        }
      },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c85"
          },
          "providerId": {
            "type": "string",
            "example": "local_60d21b4667d0d8992e610c85"
          },
          "name": {
            "type": "string",
            "example": "Jean Dupont"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jean.dupont@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+33612345678"
          },
          "emailVerified": {
            "type": "boolean",
            "example": false
          },
          "fcmTokens": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["d6Vw5PRiQ8-v0TbZm_sj2N:APA91..."]
          },
          "notificationPreferences": {
            "type": "object",
            "properties": {
              "email": {
                "type": "boolean",
                "example": true
              },
              "sms": {
                "type": "boolean",
                "example": true
              },
              "push": {
                "type": "boolean",
                "example": true
              }
            }
          },
          "active": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-05-01T12:00:00Z"
          }
        }
      },
      "Notification": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c86"
          },
          "user": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "60d21b4667d0d8992e610c85"
              },
              "name": {
                "type": "string",
                "example": "Jean Dupont"
              },
              "email": {
                "type": "string",
                "example": "jean.dupont@example.com"
              },
              "phone": {
                "type": "string",
                "example": "+33612345678"
              }
            }
          },
          "type": {
            "type": "string",
            "enum": ["email", "sms", "push"],
            "example": "email"
          },
          "title": {
            "type": "string",
            "example": "Bienvenue sur RoadTrip!"
          },
          "message": {
            "type": "string",
            "example": "Merci de vous être inscrit sur notre plateforme."
          },
          "sentAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-05-01T12:00:00Z"
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-05-01T12:00:10Z"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-05-01T12:05:00Z"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "sent", "delivered", "failed", "read"],
            "example": "sent"
          },
          "failureReason": {
            "type": "string",
            "example": "Adresse email invalide"
          },
          "metadata": {
            "type": "object",
            "example": {
              "to": "jean.dupont@example.com",
              "template": "welcome",
              "messageId": "<1234567890@example.com>"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "Une erreur s'est produite"
          }
        }
      }
    }
  }
}